finished README and renamed footer to hopefully fix Render issue.
